+++ src/main/java/com/iexec/resultproxy/challenge/ChallengeService.java
@@ -1,19 +1,32 @@
+/*
+ * Copyright 2020-2023 IEXEC BLOCKCHAIN TECH
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package com.iexec.resultproxy.challenge;
 
-import com.iexec.common.chain.eip712.entity.EIP712Challenge;
 import com.iexec.common.security.SignedChallenge;
-import com.iexec.common.utils.BytesUtils;
-import com.iexec.common.utils.SignatureUtils;
-
+import com.iexec.commons.poco.eip712.entity.EIP712Challenge;
+import com.iexec.commons.poco.utils.BytesUtils;
+import com.iexec.commons.poco.utils.SignatureUtils;
+import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.stereotype.Service;
 import org.web3j.utils.Numeric;
 
-import lombok.extern.slf4j.Slf4j;
-
-
-@Service
 @Slf4j
+@Service
 public class ChallengeService {
 
     private final EIP712ChallengeService eip712ChallengeService;